requirements.txt
boto3==1.28.84
botocore==1.31.84
Django==4.2.0
gunicorn==20.1.0
django-browser-reload==1.12.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-htmx==1.17.0
django-phonenumber-field==7.2.0
django-storages==1.14.2
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
icecream==2.1.3
phonenumbers==8.13.23
Pillow==10.1.0
psycopg2-binary
python-dotenv==1.0.0
whitenoise==6.2.0
pandas==2.0.3
plotly==5.18.0

